Split lammps data tests into their own file (#466)
